+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: lda-cunh <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 16:10:59 by lda-cunh          #+#    #+#             */
-/*   Updated: 2022/10/27 17:46:18 by lda-cunh         ###   ########.fr       */
+/*   Updated: 2023/01/07 20:08:45 by lda-cunh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ char	*ft_substr(const char *s, unsigned int start, size_t len)
 	i = 0;
 	if (!s)
 		return (0);
-	if (ft_strlen((char *)s) < start)
+	if ((unsigned int)ft_strlen((char *)s) < start)
 		return (ft_strdup(""));
-	if (ft_strlen((char *)s + start) < len)
+	if ((size_t)ft_strlen((char *)s + start) < len)
 		substr = (char *)malloc(ft_strlen((char *)s + start) + 1);
 	else
 		substr = (char *)malloc(sizeof(char) * len + 1);
